refactor: tests with helper to read xml and mocks organization
