+++ test/__mocks__/tasks/flash.xml
