+++ test/__mocks__/tasks/flash_profile_variables.xml
