+++ test/test_task_xml.py
@@ -1,23 +1,16 @@
-from lxml import etree
 from pytest import fixture
 
 from tasker.py.profile_variable import ProfileVariable
 from tasker.xml import TaskerXml
 
-
-def xml_to_string(element_xml):
-    etree.tostring(element_xml, pretty_print=True, encoding='utf-8').decode()
-
-
-def read_xml(path):
-    parser = etree.XMLParser(remove_blank_text=True)
-    mock_tree = etree.parse(path, parser)
-
-    return xml_to_string(mock_tree)
+from test.helpers.xml_reader import (
+    read_xml_task,
+    xml_to_string
+)
 
 
 def test_beep_xml(beep_task):
-    expected_xml = read_xml('test/__mocks__/beep.xml')
+    expected_xml = read_xml_task('beep.xml')
 
     task_xml = TaskerXml(tasks=[beep_task]).to_xml()
     xml_string = xml_to_string(task_xml)
@@ -25,7 +18,7 @@ def test_beep_xml(beep_task):
 
 
 def test_flash_xml(flash_task):
-    expected_xml = read_xml('test/__mocks__/flash.xml')
+    expected_xml = read_xml_task('flash.xml')
 
     task_xml = TaskerXml(tasks=[flash_task]).to_xml()
     xml_string = xml_to_string(task_xml)
@@ -33,7 +26,7 @@ def test_flash_xml(flash_task):
 
 
 def test_task_profile_variable(flash_task):
-    expected_xml = read_xml('test/__mocks__/flash_profile_variables.xml')
+    expected_xml = read_xml_task('flash_profile_variables.xml')
 
     flash_task.profile_variables = [
         ProfileVariable(variable_name='%test_variable', value='TEST_VALUE'),
